requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
